# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart LPR API",
        "description": "Open API for Smart LPR",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/history/{date_start}/{date_end}/{gate}/{page}/{limit}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get history of items",
                "description": "Retrieve a history of items based on date range, gate, pagination, and limit",
                "operationId": "history",
                "parameters": [
                    {
                        "name": "date_start",
                        "in": "path",
                        "description": "Start date for filtering the list",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "datetime",
                            "example": "2024-09-01 00:00:00"
                        }
                    },
                    {
                        "name": "date_end",
                        "in": "path",
                        "description": "End date for filtering the list",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "datetime",
                            "example": "2024-09-05 23:59:59"
                        }
                    },
                    {
                        "name": "gate",
                        "in": "path",
                        "description": "Gate number for filtering the list",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "66d9965ec312805ebe04d8bd"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List retrieved successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/v1/list/{page}/{limit}": {
            "get": {
                "tags": [
                    "Lists"
                ],
                "summary": "Get history of items",
                "description": "Retrieve a list of items based on date range, gate, pagination, and limit",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List retrieved successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/list": {
            "post": {
                "tags": [
                    "Lists"
                ],
                "summary": "Create a new list with dynamic fields",
                "operationId": "createLists",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "description": "Full name of the person",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "lpr": {
                                        "description": "License plate number",
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "note": {
                                        "description": "Additional notes",
                                        "type": "string",
                                        "example": "Some notes"
                                    },
                                    "type_list_id": {
                                        "description": "List type id",
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "start_date": {
                                        "description": "Start date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-01"
                                    },
                                    "end_date": {
                                        "description": "End date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/list/{id}": {
            "get": {
                "tags": [
                    "Lists"
                ],
                "summary": "Get list of lists",
                "description": "Returns list of lists",
                "operationId": "getListsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the list to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Message"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lists"
                ],
                "summary": "Delete a list",
                "description": "Delete a list by its ID",
                "operationId": "deleteLists",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the list to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List deleted successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "List not found"
                    }
                }
            }
        },
        "/api/v1/list/update": {
            "post": {
                "tags": [
                    "Lists"
                ],
                "summary": "Create a new list with dynamic fields",
                "operationId": "updateLists",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID of the list",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "fullname": {
                                        "description": "Full name of the person",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "lpr": {
                                        "description": "License plate number",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "note": {
                                        "description": "Additional notes",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "type_list_id": {
                                        "description": "List type id",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "start_date": {
                                        "description": "Start date",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "end_date": {
                                        "description": "end date",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object",
                                "additionalProperties": {
                                    "description": "Any additional fields provided dynamically",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}